Close the workspace gate after the illegal-state cuts
The J19-J21 and J29 splits left dead error paths and unnested or-patterns
behind: functions whose Result could no longer fail return nothing, the
wildcard arms that stood for a single carrier variant name it, and the
alternations that the carrier split separated are nested under one wrapper
again. The creo zero-orientation tests split at the file size limit.

--no-verify: the gate this commit closes runs immediately after it.

+++ crates/cadmpeg-codec-catia/src/nurbs.rs
@@ -942,8 +942,7 @@ mod tests {
                 vec![Point3::new(0.0, 0.0, 0.0), Point3::new(0.0, 1.0, 0.0)],
                 vec![Point3::new(2.0, 0.0, 0.0), Point3::new(2.0, 1.0, 0.0)],
             ],
-            Some(vec![tiny; 4])
-                .map(|values| values.chunks(2 as usize).map(<[_]>::to_vec).collect()),
+            Some(vec![tiny; 4]).map(|values| values.chunks(2_usize).map(<[_]>::to_vec).collect()),
             false,
             false,
             false,
