+++ crates/cadmpeg-codec-f3d/src/writer/tests/source_less_nurbs.rs
@@ -141,7 +141,7 @@ fn generated_source_less_face_writes_rational_nurbs_surface_carrier() {
                 ],
             ],
             Some(vec![1.0, 0.75, 1.25, 1.0])
-                .map(|values| values.chunks(2 as usize).map(<[_]>::to_vec).collect()),
+                .map(|values| values.chunks(2_usize).map(<[_]>::to_vec).collect()),
             false,
             false,
             true,
@@ -597,7 +597,7 @@ fn generated_source_less_multi_face_writes_nurbs_carriers_and_pcurve() {
                 ],
             ],
             Some(vec![1.0, 0.8, 1.2, 1.0])
-                .map(|values| values.chunks(2 as usize).map(<[_]>::to_vec).collect()),
+                .map(|values| values.chunks(2_usize).map(<[_]>::to_vec).collect()),
             false,
             false,
             true,
