+++ crates/cadmpeg-codec-rhino/src/writer/tests/mod.rs
@@ -786,7 +786,7 @@ pub(crate) fn rectangular_nurbs_patch() -> CadIr {
             vec![7.0, 7.0, 11.0, 11.0],
             vec![vec![points[0], points[3]], vec![points[1], points[2]]],
             Some(vec![1.0, 0.8, 1.2, 1.0])
-                .map(|values| values.chunks(2 as usize).map(<[_]>::to_vec).collect()),
+                .map(|values| values.chunks(2_usize).map(<[_]>::to_vec).collect()),
             false,
             false,
             false,
@@ -883,7 +883,7 @@ pub(crate) fn mixed_plane_nurbs_sheet() -> CadIr {
             vec![7.0, 7.0, 11.0, 11.0],
             vec![vec![points[0], points[3]], vec![points[1], points[2]]],
             Some(vec![1.0, 0.8, 1.2, 1.0])
-                .map(|values| values.chunks(2 as usize).map(<[_]>::to_vec).collect()),
+                .map(|values| values.chunks(2_usize).map(<[_]>::to_vec).collect()),
             false,
             false,
             false,
